examples/bridge_to_javascript/requirements.txt
fastapi>=0.109.0,<0.110.0
uvicorn>=0.27.0,<0.28.0
httpx>=0.26.0,<0.27.0
python-dotenv>=1.0.0,<2.0.0
SQLAlchemy>=1.4.0
pydantic>=2.0.0
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
async-timeout>=4.0.0,<5.0.0
PyYAML>=6.0.1,<7.0.0
